# frontend_contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Production RAG API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/organizations/current": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get Current Organization",
        "operationId": "get_current_organization_api_v1_organizations_current_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/organizations/current/members": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "List Members",
        "operationId": "list_members_api_v1_organizations_current_members_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationMemberListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create Member",
        "operationId": "create_member_api_v1_organizations_current_members_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationMemberResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/organizations/current/members/{membership_id}/role": {
      "patch": {
        "tags": [
          "Organizations"
        ],
        "summary": "Update Member Role",
        "operationId": "update_member_role_api_v1_organizations_current_members__membership_id__role_patch",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "membership_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Membership Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationMemberRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationMemberResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/current/members/{membership_id}": {
      "delete": {
        "tags": [
          "Organizations"
        ],
        "summary": "Deactivate Member",
        "operationId": "deactivate_member_api_v1_organizations_current_members__membership_id__delete",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "membership_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Membership Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOrganizationMemberRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password"
          },
          "role": {
            "$ref": "#/components/schemas/OrganizationRole",
            "default": "member"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email",
          "password"
        ],
        "title": "CreateOrganizationMemberRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrganizationMemberListResponse": {
        "properties": {
          "members": {
            "items": {
              "$ref": "#/components/schemas/OrganizationMemberResponse"
            },
            "type": "array",
            "title": "Members"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "members",
          "total"
        ],
        "title": "OrganizationMemberListResponse"
      },
      "OrganizationMemberResponse": {
        "properties": {
          "membership_id": {
            "type": "string",
            "title": "Membership Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/OrganizationRole"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "membership_id",
          "user_id",
          "email",
          "full_name",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "OrganizationMemberResponse"
      },
      "OrganizationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "current_role": {
            "$ref": "#/components/schemas/OrganizationRole"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "current_role"
        ],
        "title": "OrganizationResponse"
      },
      "OrganizationRole": {
        "type": "string",
        "enum": [
          "owner",
          "admin",
          "member",
          "viewer"
        ],
        "title": "OrganizationRole",
        "description": "Roles available inside an organization."
      },
...
      "UpdateOrganizationMemberRoleRequest": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/OrganizationRole"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "UpdateOrganizationMemberRoleRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "Enter a signed JWT access token.",
        "scheme": "bearer"
      }
    }
  }
}